refactor: Move the snap interface to a separate file
